// src/services/paths.d.ts
import type { SoilDatabase } from "..";
type DataType = keyof SoilDatabase;
type DataKey = string | number | symbol;
export declare const PATHS: {
    ADMINS: string;
    admin: (uid: string) => string;
    USERS: string;
    user: (uid: string) => string;
    TRACKING: string;
    trackingKey: (trackingKey: string) => string;
    DATA: string;
    dataType: (dataType: DataType) => string;
    dataKey: (dataType: DataType, dataKey: DataKey) => string;
    dataKeyField: (dataType: DataType, dataKey: DataKey, dataField: DataKey) => string;
    OWNERS: string;
    ownerDataType: (dataType: DataType) => string;
    ownerDataKey: (dataType: DataType, dataKey: DataKey) => string;
    ownerDataKeyUid: (dataType: DataType, dataKey: DataKey, uid: string) => string;
    USER_DATA_LISTS: string;
    userDataList: (uid: string) => string;
    userDataTypeList: (uid: string, dataType: DataType) => string;
    userDataKeyList: (uid: string, dataType: DataType, dataKey: DataKey) => string;
    PUBLIC_DATA_LISTS: string;
    publicDataTypeList: (dataType: DataType) => string;
    publicDataKeyList: (dataType: DataType, dataKey: DataKey) => string;
    CONNECTION_DATA_LISTS: string;
    connectionDataListType: (dataType: DataType) => string;
    connectionDataListKey: (dataType: DataType, dataKey: DataKey) => string;
    connectionDataListConnectionType: (dataType: DataType, dataKey: DataKey, connectionType: DataType) => string;
    connectionDataListConnectionKey: (dataType: DataType, dataKey: DataKey, connectionType: DataType, connectionKey: DataKey) => string;
};
export declare const DB_DELIMITER = "__";
export declare const generateDbKey: (...keys: string[]) => string;
export declare const parseDbKey: (dbKey: string) => string[];
export declare const cleanPushKey: (key: string) => string;
export {};